Отрисуй меня полностью (часть 2) fix
